# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overtime Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Roster {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface RosterInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface RosterSettings {
  rosterId: number;
  useOfferedHours: boolean;
  useSeniority: boolean;
  useSubclassOrdering: boolean;
}

export interface RosterSettingsInput {
  useOfferedHours?: boolean;
  useSeniority?: boolean;
  useSubclassOrdering?: boolean;
}

export interface DayTypeConfig {
  rosterId: number;
  dayType: string;
  name: string;
  enabled: boolean;
  /** @nullable */
  multiplier?: number | null;
  sortOrder: number;
}

export interface DayTypeConfigInput {
  name?: string;
  enabled?: boolean;
  /** @nullable */
  multiplier?: number | null;
  sortOrder?: number;
}

export interface CreateDayTypeConfigInput {
  dayType: string;
  name: string;
  enabled?: boolean;
  /** @nullable */
  multiplier?: number | null;
  sortOrder?: number;
}

export interface SubclassDayTypeSortEntry {
  subclassId: number;
  name: string;
  /** Effective sort order for this day type (override if isOverride, else global) */
  sortOrder: number;
  /** True when a day-type-specific override exists for this subclass */
  isOverride: boolean;
}

export interface SubclassDayTypeSortInputItem {
  subclassId: number;
  sortOrder: number;
}

export interface Role {
  id: number;
  rosterId: number;
  name: string;
}

export interface RoleInput {
  /** @minLength 1 */
  name: string;
}

export interface Subclass {
  id: number;
  rosterId: number;
  name: string;
  /** Lower value = higher priority in Up Next ordering */
  sortOrder: number;
}

export interface SubclassInput {
  /** @minLength 1 */
  name: string;
  sortOrder?: number;
}

export interface Employee {
  id: number;
  rosterId: number;
  name: string;
  seniority: number;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  roleName?: string | null;
  /** @nullable */
  subclassId?: number | null;
  /** @nullable */
  subclassName?: string | null;
  active: boolean;
  totalOfferedHours?: number;
  totalWorkedHours?: number;
  /** Weighted offered hours used for rotation ordering */
  fairnessScore?: number;
  /** Baseline hours used for fairness computation; not counted in offered/worked totals */
  startingNormalizedHours?: number;
}

export interface EmployeeInput {
  rosterId: number;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  seniority: number;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  subclassId?: number | null;
  active?: boolean;
  /** @minimum 0 */
  startingNormalizedHours?: number;
}

export interface EmployeeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 1 */
  seniority?: number;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  subclassId?: number | null;
  active?: boolean;
  /** @minimum 0 */
  startingNormalizedHours?: number;
}

export interface EventEntry {
  id: number;
  employeeId: number;
  employeeName: string;
  offered: boolean;
  worked: boolean;
  /** @nullable */
  hoursOverride?: number | null;
  hoursAwarded: number;
  hoursOffered: number;
}

export interface Event {
  id: number;
  rosterId: number;
  date: string;
  description: string;
  defaultHours: number;
  dayType: string;
  multiplier: number;
  entries: EventEntry[];
}

export interface EventEntryInput {
  employeeId: number;
  offered: boolean;
  worked: boolean;
  /** @nullable */
  hoursOverride?: number | null;
}

export interface EventInput {
  rosterId: number;
  date: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  defaultHours: number;
  dayType: string;
  /** @minimum 0 */
  multiplier?: number;
  /** @minItems 1 */
  entries: EventEntryInput[];
}

export interface UpNextEmployee {
  id: number;
  name: string;
  /** @nullable */
  subclassId?: number | null;
  /** @nullable */
  subclassName?: string | null;
  /** @nullable */
  roleName?: string | null;
  seniority: number;
  totalOfferedHours: number;
  fairnessScore: number;
  rank: number;
}

export interface UpNextResult {
  rosterId: number;
  dayType: string;
  employees: UpNextEmployee[];
}

export interface EmployeeEventRecord {
  eventId: number;
  date: string;
  description: string;
  dayType: string;
  offered: boolean;
  worked: boolean;
  hoursOffered: number;
  hoursAwarded: number;
  /** @nullable */
  hoursOverride?: number | null;
  multiplier: number;
}

export interface EmployeeReport {
  employee: Employee;
  totalOfferedHours: number;
  totalWorkedHours: number;
  acceptanceRate?: number;
  events: EmployeeEventRecord[];
}

export interface EmployeeStat {
  id: number;
  name: string;
  totalWorkedHours: number;
  totalOfferedHours: number;
}

export interface Stats {
  totalEvents: number;
  totalOfferedHours: number;
  totalWorkedHours: number;
  employeeCount: number;
  topWorkers?: EmployeeStat[];
  recentEvents?: Event[];
  /**
     * ID of the most recent "Reset Hours" event used for time-window filtering
     * @nullable
     */
  normalizedHoursResetEventId?: number | null;
}

export type DayTypeSuggestionSuggestedDayType = typeof DayTypeSuggestionSuggestedDayType[keyof typeof DayTypeSuggestionSuggestedDayType];


export const DayTypeSuggestionSuggestedDayType = {
  weekday: 'weekday',
  weekend: 'weekend',
  holiday: 'holiday',
} as const;

export interface DayTypeSuggestion {
  date: string;
  suggestedDayType: DayTypeSuggestionSuggestedDayType;
  reason: string;
  isHoliday?: boolean;
  /** @nullable */
  holidayName?: string | null;
}

export type ListEmployeesParams = {
rosterId?: number;
};

export type ListEventsParams = {
rosterId?: number;
};

export type GetUpNextParams = {
rosterId: number;
dayType?: string;
};

export type GetStatsParams = {
rosterId?: number;
};

export type SuggestDayTypeParams = {
date: string;
};