# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overtime Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rosters/{id}/subclass-day-type-sort/{dayType}:
    get:
      operationId: listSubclassDayTypeSort
      tags: [classifications]
      summary: List subclasses ordered by day-type-specific sort, falling back to global
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: dayType
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ordered subclass list for this day type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubclassDayTypeSortEntry"
    put:
      operationId: putSubclassDayTypeSort
      tags: [classifications]
      summary: Bulk-set per-day-type subclass order (replaces all rows for this day type)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: dayType
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SubclassDayTypeSortInputItem"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubclassDayTypeSortEntry"
    delete:
      operationId: deleteSubclassDayTypeSort
      tags: [classifications]
      summary: Reset per-day-type subclass order (reverts to global default)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: dayType
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Roles ─────────────────────────────────────────────────────────────────
  /rosters/{rosterId}/roles:
    get:
      operationId: listRoles
      tags: [classifications]
      summary: List roles for a roster
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [classifications]
      summary: Create a role
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /rosters/{rosterId}/roles/{id}:
    patch:
      operationId: updateRole
      tags: [classifications]
      summary: Update a role
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Not found
    delete:
      operationId: deleteRole
      tags: [classifications]
      summary: Delete a role
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Subclasses ────────────────────────────────────────────────────────────
  /rosters/{rosterId}/subclasses:
    get:
      operationId: listSubclasses
      tags: [classifications]
      summary: List subclasses for a roster
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subclasses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subclass"
    post:
      operationId: createSubclass
      tags: [classifications]
      summary: Create a subclass
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubclassInput"
      responses:
        "201":
          description: Created subclass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subclass"
  /rosters/{rosterId}/subclasses/{id}:
    patch:
      operationId: updateSubclass
      tags: [classifications]
      summary: Update a subclass
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubclassInput"
      responses:
        "200":
          description: Updated subclass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subclass"
        "404":
          description: Not found
    delete:
      operationId: deleteSubclass
      tags: [classifications]
      summary: Delete a subclass
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Employees ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SubclassDayTypeSortEntry:
      type: object
      required: [subclassId, name, sortOrder, isOverride]
      properties:
        subclassId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
          description: Effective sort order for this day type (override if isOverride, else global)
        isOverride:
          type: boolean
          description: True when a day-type-specific override exists for this subclass
    SubclassDayTypeSortInputItem:
      type: object
      required: [subclassId, sortOrder]
      properties:
        subclassId:
          type: integer
        sortOrder:
          type: integer
    Role:
      type: object
      required: [id, rosterId, name]
      properties:
        id:
          type: integer
        rosterId:
          type: integer
        name:
          type: string
    RoleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    Subclass:
      type: object
      required: [id, rosterId, name, sortOrder]
      properties:
        id:
          type: integer
        rosterId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
          description: Lower value = higher priority in Up Next ordering
    SubclassInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer

    # ── Employees ─────────────────────────────────────────────────────────────
...